* master: (1641 commits)
  Translate comment in test
  Added test to ANTLR skip_list.json
  Update Client.cpp
  Fixed race in tests
  Update 01014_lazy_database_basic.sh
  Update 00459_group_array_insert_at.sql
  Fixed test_kafka_flush_by_time test.
  LowCardinality UUID fix
  Same for ALTER
  Update mergetree.md
  Syntax and links
  Added test.
  Check where and prewhere identifiers exist.
  Clickhouse client query param CTE added test
  Remove redundant lsof
  correct style zh doc typo
  Add lsof to fasttest
  Add new line to status
  Make integration odbc tests idempotent
  Minor code improvement in JOIN
  ...

# Conflicts:
#	src/Access/Authentication.cpp
#	src/Access/Authentication.h
#	src/Access/ExternalAuthenticators.cpp
#	src/Access/ExternalAuthenticators.h
#	src/Access/IAccessStorage.cpp
#	src/Access/LDAPAccessStorage.cpp
#	src/Access/LDAPAccessStorage.h
#	src/Access/LDAPParams.h
#	src/Common/ErrorCodes.cpp
#	src/Core/config_core.h.in

+++ tests/testflows/ldap/authentication/tests/common.py
@@ -78,7 +78,7 @@ def restart(node=None, safe=False, timeout=60):
                     f"ConfigReloader: Loaded config '/etc/clickhouse-server/config.xml', performed update on configuration",
                     timeout=timeout)
 
-def add_config(config, timeout=60, restart=False):
+def add_config(config, timeout=60, restart=False, modify=False):
     """Add dynamic configuration file to ClickHouse.
 
     :param node: node
@@ -165,19 +165,20 @@ def add_config(config, timeout=60, restart=False):
                     wait_for_config_to_be_loaded()
         yield
     finally:
-        with Finally(f"I remove {config.name}"):
-            with node.cluster.shell(node.name) as bash:
-                bash.expect(bash.prompt)
-                bash.send("tail -n 0 -f /var/log/clickhouse-server/clickhouse-server.log")
+        if not modify:
+            with Finally(f"I remove {config.name}"):
+                with node.cluster.shell(node.name) as bash:
+                    bash.expect(bash.prompt)
+                    bash.send("tail -n 0 -f /var/log/clickhouse-server/clickhouse-server.log")
 
-                with By("removing the config file", description=config.path):
-                    node.command(f"rm -rf {config.path}", exitcode=0)
+                    with By("removing the config file", description=config.path):
+                        node.command(f"rm -rf {config.path}", exitcode=0)
 
-                with Then(f"{config.preprocessed_name} should be updated", description=f"timeout {timeout}"):
-                    check_preprocessed_config_is_updated(after_removal=True)
+                    with Then(f"{config.preprocessed_name} should be updated", description=f"timeout {timeout}"):
+                        check_preprocessed_config_is_updated(after_removal=True)
 
-                with And("I wait for config to be reloaded"):
-                    wait_for_config_to_be_loaded()
+                    with And("I wait for config to be reloaded"):
+                        wait_for_config_to_be_loaded()
 
 def create_ldap_servers_config_content(servers, config_d_dir="/etc/clickhouse-server/config.d", config_file="ldap_servers.xml"):
     """Create LDAP servers configuration content.
@@ -202,11 +203,18 @@ def create_ldap_servers_config_content(servers, config_d_dir="/etc/clickhouse-se
     return Config(content, path, name, uid, "config.xml")
 
 @contextmanager
+def modify_config(config, restart=False):
+    """Apply updated configuration file.
+    """
+    return add_config(config, restart=restart, modify=True)
+
+@contextmanager
 def ldap_servers(servers, config_d_dir="/etc/clickhouse-server/config.d", config_file="ldap_servers.xml",
-        timeout=60, restart=False):
+        timeout=60, restart=False, config=None):
     """Add LDAP servers configuration.
     """
-    config = create_ldap_servers_config_content(servers, config_d_dir, config_file)
+    if config is None:
+        config = create_ldap_servers_config_content(servers, config_d_dir, config_file)
     return add_config(config, restart=restart)
 
 def create_ldap_users_config_content(*users, config_d_dir="/etc/clickhouse-server/users.d", config_file="ldap_users.xml"):
